+++ src/lib/config/Options.ts
@@ -266,7 +266,8 @@ class Options {
             sendMessageToConnectedUser: false,
             conversationsRetrievedFormat: "small",
             storeMessages: false,
-            copyMessage: false
+            copyMessage: false,
+            nbMaxConversations: 15
         };
 
         if (!("sendReadReceipt" in this._options.im)) {
@@ -281,6 +282,7 @@ class Options {
         optionsIM.conversationsRetrievedFormat = this._options.im.conversationsRetrievedFormat ? this._options.im.conversationsRetrievedFormat : config.im.conversationsRetrievedFormat;
         optionsIM.storeMessages = this._options.im.storeMessages ? this._options.im.storeMessages : config.im.storeMessages;
         optionsIM.copyMessage = this._options.im.copyMessage ? this._options.im.copyMessage : config.im.copyMessage;
+        optionsIM.nbMaxConversations = this._options.im.nbMaxConversations ? this._options.im.nbMaxConversations : config.im.nbMaxConversations;
 
         return optionsIM;
     }